+++ app/src/main/java/com/weiting/tohealth/data/FirebaseRepository.kt
@@ -135,7 +135,7 @@ interface FirebaseRepository {
 
     suspend fun getUserInfo(userId: String): User
 
-    fun editStock(itemId: String, num: Int)
+    fun editStock(itemId: String, num: Float)
 
     fun postNotification(notification: Notification)
